Rename color-rgba -> ColorRGBA
